add: change topic verification, delete of componentes where change topic, section formatation and consult of the aula right

+++ componente.php
@@ -84,6 +84,7 @@ if ($_SERVER['REQUEST_METHOD'] == 'GET') {
 
   // Formata os dados recebidos por post e get
   $cms = [];
+  $total = 0;
   foreach ($_POST as $post => $value) {
     if(substr($post, 0, 10) == "componente") {
       $cmid = substr($post, 11);
@@ -91,37 +92,44 @@ if ($_SERVER['REQUEST_METHOD'] == 'GET') {
       $cms[$cmid]->aulaid = filter_input(INPUT_GET, 'aulaid', FILTER_VALIDATE_INT);
       $cms[$cmid]->moduleid = substr($post, 11);
       $cms[$cmid]->quantidade_aulas = $_POST["presenca-{$cmid}"];
+      $total += $_POST["presenca-{$cmid}"];
     }
   }
 
-  // Compara os dados vindo do banco com o que veio do POST e se ouver alteração, altera. Além disso, se o objeto já existir no banco, retira dos arrays.
-  // Sobrando do array que veio do banco, os componentes que precisam ser apagados e do array que veio do post os componentes que precisam ser adicionados
-  foreach ($componentes as $comp => $value) {
-    if (array_key_exists($value->moduleid, $cms)) {
-      if ($cms[$value->moduleid]->quantidade_aulas != $value->quantidade_aulas) {
-        $cms[$value->moduleid]->id = $value->id;
-        $DB->update_record('suapattendance_componente', $cms[$value->moduleid]);
+  if ($total == 100) {
+    // Compara os dados vindo do banco com o que veio do POST e se ouver alteração, altera. Além disso, se o objeto já existir no banco, retira dos arrays.
+    // Sobrando do array que veio do banco, os componentes que precisam ser apagados e do array que veio do post os componentes que precisam ser adicionados
+    foreach ($componentes as $comp => $value) {
+      if (array_key_exists($value->moduleid, $cms)) {
+        if ($cms[$value->moduleid]->quantidade_aulas != $value->quantidade_aulas) {
+          $cms[$value->moduleid]->id = $value->id;
+          $DB->update_record('suapattendance_componente', $cms[$value->moduleid]);
+        }
+        unset($componentes[$comp]);
+        unset($cms[$value->moduleid]);
       }
-      unset($componentes[$comp]);
-      unset($cms[$value->moduleid]);
     }
+    
+    // Adicionando no banco os novos componentes
+    if ($cms) {
+      foreach ($cms as $cm) {
+        $DB->insert_record('suapattendance_componente', $cm, $returnid=false, $bulk=false);
+      }
+    }
+
+    // Apagando do banco componentes não mais utilizados
+    if ($componentes) {
+      foreach($componentes as $comp) {
+        $DB->delete_records('suapattendance_componente', ['id'=>$comp->id]);
+      }
+    }
+
+    redirect("{$CFG->wwwroot}/blocks/suapattendance/configurar-frequencia.php?courseid=$COURSE->id", "Presença configurada com sucesso!");
+  } else {
+    redirect("{$CFG->wwwroot}/blocks/suapattendance/configurar-frequencia.php?courseid=$COURSE->id");
   }
+
   
-  // Adicionando no banco os novos componentes
-  if ($cms) {
-    foreach ($cms as $cm) {
-      $DB->insert_record('suapattendance_componente', $cm, $returnid=false, $bulk=false);
-    }
-  }
-
-  // Apagando do banco componentes não mais utilizados
-  if ($componentes) {
-    foreach($componentes as $comp) {
-      $DB->delete_records('suapattendance_componente', ['id'=>$comp->id]);
-    }
-  }
-
-  redirect("{$CFG->wwwroot}/blocks/suapattendance/configurar-frequencia.php?courseid=$COURSE->id", "Presença configurada com sucesso!");
 }
 
 echo $OUTPUT->footer();
